application/requirements.txt
anthropic>=0.25.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
python-docx>=1.0.0
pytest>=8.0.0
pytest-mock>=3.0.0
